refactor: extract base group option
