+++ src/commands/group/settings/SpecificGroupCommandSettings.cs
@@ -1,10 +1,11 @@
 using System.ComponentModel;
+using SimpleGitShell.Commands.Base.Settings;
 using SimpleGitShell.Library.Utils;
 using Spectre.Console.Cli;
 
 namespace SimpleGitShell.Commands.Group.Settings;
 
-public class SpecificGroupCommandSettings : BaseGroupCommandSettings
+public class SpecificGroupCommandSettings : BaseGroupSettings
 {
     [Description("The name of the group.")]
     [CommandArgument(0, "<group>")]
