+++ src/commands/ssh/user/settings/BaseSSHCommandSettings.cs
